change data type in mask_blob of pooling from int64_t to uint32_t.

+++ oneflow/core/kernel/pooling_kernel.cu
@@ -9,7 +9,7 @@ namespace {
 template<typename FloatingPointType>
 __global__ void MaxPoolForward(const int64_t nthreads,
                                const FloatingPointType* in_dptr,
-                               FloatingPointType* out_dptr, int64_t* mask_dptr,
+                               FloatingPointType* out_dptr, uint32_t* mask_dptr,
                                const int64_t channels, const int64_t height,
                                const int64_t width, const int64_t pooled_height,
                                const int64_t pooled_width,
@@ -32,7 +32,7 @@ __global__ void MaxPoolForward(const int64_t nthreads,
     const FloatingPointType* const in_slice =
         in_dptr + (n * channels + c) * height * width;
     FloatingPointType maxval = in_slice[hstart * width + wstart];
-    int64_t maxidx = hstart * width + wstart;
+    uint32_t maxidx = hstart * width + wstart;
     for (int64_t h = hstart; h < hend; ++h) {
       for (int64_t w = wstart; w < wend; ++w) {
         if (in_slice[h * width + w] > maxval) {
@@ -87,7 +87,7 @@ __global__ void AvePoolForward(const int64_t nthreads,
 template<typename FloatingPointType>
 __global__ void MaxPoolBackward(
     const int64_t nthreads, const FloatingPointType* out_diff_dptr,
-    const int64_t* mask_dptr, FloatingPointType* in_diff_dptr,
+    const uint32_t* mask_dptr, FloatingPointType* in_diff_dptr,
     const int64_t channels, const int64_t height, const int64_t width,
     const int64_t pooled_height, const int64_t pooled_width,
     const int64_t kernel_h, const int64_t kernel_w, const int64_t stride_h,
@@ -110,7 +110,7 @@ __global__ void MaxPoolBackward(
     FloatingPointType gradient = 0;
     const int64_t offset = (n * channels + c) * pooled_height * pooled_width;
     const FloatingPointType* const out_diff_slice = out_diff_dptr + offset;
-    const int64_t* const mask_slice = mask_dptr + offset;
+    const uint32_t* const mask_slice = mask_dptr + offset;
     for (int64_t ph = phstart; ph < phend; ++ph) {
       for (int64_t pw = pwstart; pw < pwend; ++pw) {
         if (mask_slice[ph * pooled_width + pw] == h * width + w) {
@@ -181,7 +181,7 @@ class PoolingKernelUtil<DeviceType::kGPU, FloatingPointType> final {
                ctx.device_ctx->cuda_stream()>>>(
                 count, in_blob->dptr<FloatingPointType>(),
                 out_blob->mut_dptr<FloatingPointType>(),
-                mask_blob->mut_dptr<int64_t>(), in_blob->shape().At(1),
+                mask_blob->mut_dptr<uint32_t>(), in_blob->shape().At(1),
                 in_blob->shape().At(2), in_blob->shape().At(3),
                 out_blob->shape().At(2), out_blob->shape().At(3),
                 pooling_conf.kernel_size(0), pooling_conf.kernel_size(1),
@@ -220,7 +220,7 @@ class PoolingKernelUtil<DeviceType::kGPU, FloatingPointType> final {
             <<<BlocksNum4ThreadsNum(count), kCudaThreadsNumPerBlock, 0,
                ctx.device_ctx->cuda_stream()>>>(
                 count, out_diff_blob->dptr<FloatingPointType>(),
-                mask_blob->dptr<int64_t>(),
+                mask_blob->dptr<uint32_t>(),
                 in_diff_blob->mut_dptr<FloatingPointType>(),
                 in_diff_blob->shape().At(1), in_diff_blob->shape().At(2),
                 in_diff_blob->shape().At(3), out_diff_blob->shape().At(2),
